requirements.txt
pymongo
djangorestframework
cloudinary
mongoengine
django
bcrypt
python-dotenv
django-rest-framework-mongoengine
django-extensions